[FE_Misc]confirm문 복구

+++ client/src/Components/Comments.js
@@ -111,37 +111,41 @@ const Comments = ({
   const [isEditOn, setIsEditOn] = useState(false);
 
   const addComment = () => {
-    async function postComment() {
-      const response = await dispatch(
-        postCommentThunk({
-          questionId: currentQuestion.questionId,
-          text: newComment,
-          cookie: cookies.access_token,
-        })
-      ).then((res) => {
-        setNewComment('');
-        handleRender(!render);
-      });
+    if (confirm('댓글을 추가하시겠습니까?')) {
+      async function postComment() {
+        const response = await dispatch(
+          postCommentThunk({
+            questionId: currentQuestion.questionId,
+            text: newComment,
+            cookie: cookies.access_token,
+          })
+        ).then((res) => {
+          setNewComment('');
+          handleRender(!render);
+        });
+      }
+      postComment();
     }
-    postComment();
   };
 
   const deleteComment = (event) => {
-    async function deleteComment() {
-      const response = await dispatch(
-        deleteCommentThunk({
-          questionId: currentQuestion.questionId,
-          commentId: commentsData[event.target.id].commentId,
-          cookie: cookies.access_token,
-        })
-      ).then((data) => handleRender(!render));
+    if (confirm('정말 삭제하시겠습니까?')) {
+      async function deleteComment() {
+        const response = await dispatch(
+          deleteCommentThunk({
+            questionId: currentQuestion.questionId,
+            commentId: commentsData[event.target.id].commentId,
+            cookie: cookies.access_token,
+          })
+        ).then((data) => handleRender(!render));
+      }
+      deleteComment();
+      setCommentsData(
+        commentsData.filter(
+          (a) => a.commentId !== commentsData[event.target.id].commentId
+        )
+      );
     }
-    deleteComment();
-    setCommentsData(
-      commentsData.filter(
-        (a) => a.commentId !== commentsData[event.target.id].commentId
-      )
-    );
   };
 
   const toggleEdit = (event) => {
@@ -151,21 +155,23 @@ const Comments = ({
   };
 
   const editComment = () => {
-    async function editComment() {
-      const response = await dispatch(
-        patchCommentThunk({
-          questionId: currentQuestion.questionId,
-          commentId: currentCommentId,
-          text: newComment,
-          cookie: cookies.access_token,
-        })
-      ).then((res) => {
-        setNewComment('');
-        handleRender(!render);
-      });
+    if (confirm('댓글을 수정하시겠습니까?')) {
+      async function editComment() {
+        const response = await dispatch(
+          patchCommentThunk({
+            questionId: currentQuestion.questionId,
+            commentId: currentCommentId,
+            text: newComment,
+            cookie: cookies.access_token,
+          })
+        ).then((res) => {
+          setNewComment('');
+          handleRender(!render);
+        });
+      }
+      editComment();
+      setIsEditOn(false);
     }
-    editComment();
-    setIsEditOn(false);
   };
 
   const cancelEdit = () => {
